backend/requirements.txt
fastapi
sqlalchemy
uvicorn
python-dotenv
psycopg2-binary
